More idiomatic handling of raw sensor data.
Sensor returns u16 as VOC or NOX measurements and it makes sense to use
one as an input to process(). This makes checking for overflow/underflow
unnecessary.

clamp() is a more idiomatic way to restrict value range.

+++ sensirion-gas-index-algorithm-rs/src/tests.rs
@@ -25,7 +25,7 @@ fn test_nox_algorithm() {
     assert_eq!(res, Some(1));
 }
 
-const DATASET: [(i32, Option<i32>); 600] = [
+const DATASET: [(u16, Option<i32>); 600] = [
     (20019, None),
     (20004, None),
     (20013, None),
